fix: 토큰 만료 변경

+++ src/main/java/com/example/oopclass/security/JwtUtils.java
@@ -18,7 +18,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${app.jwtExpirationMs}")
-    private long jwtExpirationMs;
+    private int jwtExpirationMs;
 
     private final SecretKey key;
